refactor: extract `cleanSpecifiedTail`
